Fixed issue related to dark theme in Level7
Android by default would kill the activity when it detect the theme is changing to redraw the activity in the respected theme, however this would also mean that the timer would reset because the onDestroy() method is called.

To solve this the activity have to make the activity to not call onDestroy() method by adding "android:configChanges="uiMode"" in the manifest, and the theme inside of the level have to be changed programmatically.